# swagger.yml
openapi: 3.0.0
info:
  title: API de Gestão de Provas
  description: API para criação, edição e gerenciamento de provas, questões, descritores.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /auth/register:
    post:
      tags:
        - Autenticação
      summary: Registra um novo usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Usuário registrado com sucesso.
  /auth/login:
    post:
      tags:
        - Autenticação
      summary: Realiza o login do usuário.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Login bem-sucedido.
...